+++ src/io/aviso/tracker.clj
@@ -11,9 +11,8 @@
   (:import [java.util WeakHashMap Map]))
 
 ;;; Contains a vector of messages (or functions that return messages) used to log the route to the exception.
-(def ^:private ^:dynamic *operation-labels*
-  "Contains a vector of operation labels (as passed to [[trace]]). These are used to log operations, when exceptions
-  occur."
+(def ^:no-doc ^:dynamic *operation-labels*
+  "Contains a vector of operation labels used to log operations, when exceptions occur."
   [])
 
 (def ^:dynamic *log-trace-level*
@@ -29,7 +28,8 @@
 (def ^:private ^Map log-exceptions?
   (WeakHashMap.))
 
-(defn- set-should-log-operations!
+;; made public for a macro, but is internal and not documented
+(defn ^:no-doc set-should-log-operations!
   [flag]
   (locking log-exceptions?
     (if flag
@@ -41,7 +41,7 @@
   (locking log-exceptions?
     (.get log-exceptions? (Thread/currentThread))))
 
-(defn- label-to-string
+(defn ^:no-doc label-to-string
   [message]
   (str (if (fn? message) (message) message)))
 
@@ -55,7 +55,8 @@
 (defn- error [logger message]
   (l/log* logger :error nil message))
 
-(defn- log-trace-label-stack [logger trace-labels e]
+(defn- log-trace-label-stack
+  [logger trace-labels e]
   (let [lines (concat ["An exception has occurred:"]
                       (map-indexed (fn [i label]
                                      (format "[%3d] - %s" (inc i) label))
@@ -65,36 +66,39 @@
          (str/join writer/eol)
          (error logger))))
 
-(defn checkpoint*
+(defn ^:no-doc handle-checkpoint-exception
+  "Handler for exceptions inside [[checkpoint*]]."
+  [logger t]
+  (if (should-log-operations?)
+    (let [trace-label-strings (operation-labels)]
+      (log-trace-label-stack logger trace-label-strings t)
+      (set-should-log-operations! false)
+      (throw (ex-info (to-message t)
+                      {:operation-trace trace-label-strings}
+                      t)))
+    (throw t)))
+
+(defmacro checkpoint*
   "The main implementation of [[checkpoint]], using a specific logger.
 
   :logger
   : logger to use when logging the operation trace.
 
-  :f
-  : function to evaluate.
+  :body
+  : forms to evaluate
 
-  The result of the function is returned.
-
-  If the function throws an exception, it may be logged (and rethrown).  When the exception is logged,
+  If the body throws an exception, it may be logged (and rethrown).  When the exception is logged,
   it includes the operation trace (the sequence of active operation labels), then the formatted exception.
 
   In the logging case, the exception is rethrown, wrapped as a new ex-info exception, with :operation-trace
   as the vector of operation label strings."
   {:added "0.1.5"}
-  [logger f]
-  (try
-    (set-should-log-operations! true)
-    (f)
-    (catch Throwable t
-      (if (should-log-operations?)
-        (let [trace-label-strings (operation-labels)]
-          (log-trace-label-stack logger trace-label-strings t)
-          (set-should-log-operations! false)
-          (throw (ex-info (to-message t)
-                          {:operation-trace trace-label-strings}
-                          t)))
-        (throw t)))))
+  [logger & body]
+  `(try
+     (set-should-log-operations! true)
+     ~@body
+     (catch Throwable t#
+       (handle-checkpoint-exception ~logger t#))))
 
 (defmacro checkpoint
   "Establishes a checkpoint, a point at which a thrown exception results in
@@ -103,13 +107,13 @@
   checkpoint is useful when flow of control passes from one thread to another, such as
   when using clojure.core.async. Although the operation trace, being stored in a dynamic var,
   will be conveyed to the new thread, the exception handler necessary to report an exception
-  is not. checkpoint provides that exception handler."
+  is not conveyed. checkpoint provides that exception handler."
   {:added "0.1.5"}
   [& body]
-  `(checkpoint* (get-logger ~*ns*) (fn [] ~@body)))
+  `(checkpoint* (get-logger ~*ns*) ~@body))
 
-(defn track*
-  "Tracks the execution of a function of no arguments. The [[track]] macro converts into a call to track*.
+(defmacro track*
+  "Tracks the execution of an expression. The [[track]] macro converts into a call to track*.
   Uses [[checkpoint]].
   
   logger
@@ -117,23 +121,23 @@
 
   label
   : String, object, or function. If a function, the function will only be evaluated if the message
-    needs to be printed (when an exception occurs, or when logging traces via [[*log-trace-level*]]).
-
-  f
-  : function to invoke. track* returns the result from this function."
-  [logger label f]
-  (let [level' *log-trace-level*
-        ;; Convert the label to a string immediately, if it is going to be logged
-        ;; immediately. Otherwise, leave the object or function as-is, until it is actually needed.
-        label' (if (and level'
-                        (fn? label)
-                        (impl/enabled? logger level'))
-                 (label-to-string label)
-                 label)]
-    (binding [*operation-labels* (conj *operation-labels* label')]
-      (when level'
-        (l/log* logger *log-trace-level* nil label'))
-      (checkpoint* logger f))))
+    needs to be printed (when an exception occurs, or when logging traces via [[*log-trace-level*]])."
+  [logger label & body]
+  (let [logger-sym (gensym "logger-")]
+    `(let [~logger-sym ~logger
+           ;; Convert the label to a string immediately, if it is going to be logged
+           ;; immediately. Otherwise, leave the object or function as-is, until it is actually needed.
+           converted-label# ~(if (string? label)
+                               label
+                               `(let [label# ~label]
+                                  (if (and *log-trace-level*
+                                           (impl/enabled? ~logger-sym *log-trace-level*))
+                                    (label-to-string label#)
+                                    label#)))]
+       (binding [*operation-labels* (conj *operation-labels* converted-label#)]
+         (when *log-trace-level*
+           (l/log* ~logger-sym *log-trace-level* nil converted-label#))
+         (checkpoint* ~logger-sym ~@body)))))
 
 (defmacro track
   "Tracks the execution of an operation, associating the label (which describes the operation) with the execution of the body.
@@ -161,7 +165,7 @@
 
   See also: [[*log-trace-level*]]."
   [label & body]
-  `(track* (get-logger ~*ns*) ~label #(do ~@body)))
+  `(track* (get-logger ~*ns*) ~label ~@body))
 
 (defn timer*
   "Executes a function, timing the duration. It then calculates the elapsed time in milliseconds (as a double) and passes that to a second function.